requirements.txt
google-generativeai==0.3.2
pdf2image==1.17.0
Pillow==10.3.0
streamlit==1.38.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pypdfium2==4.26.0
graphviz==0.20.1
plotly==5.18.0
